email_generator/requirements.txt
flask==2.3.3
pika==1.3.2
requests==2.31.0
python-dotenv==1.0.0